test(db): consolidate DB integration tests under test/integration/database
Group the three Postgres-backed integration test files (legacy OSS
bridge, migrator search_path, orchestrator) into one external package
at test/integration/database. Their near-identical per-test database
plumbing (freshDB x2 + newDB, distinct db-name prefixes) collapses into
a single freshDB in freshdb_test.go with one test_db_ prefix.

With the scattered files gone, the dbtest bridge package has no
remaining consumers outside v1alpha1store, so delete it and fold the
DSN helpers into pkg/registry/v1alpha1store/testutil.go as unexported
adminDSN/validateAdminDSN/testDBURI. The exported test surface shrinks
to NewTestPool + NewTestPoolWithDSN. The dbtest behavior pins move to
an in-package testutil_test.go.

Organization-only: assertions and exercised code paths are unchanged;
failure messages still never echo the DSN.
