+++ test/integration/database/migrate_searchpath_test.go
@@ -1,63 +1,21 @@
 //go:build integration
 
-package database_test
+package database
 
 import (
 	"context"
-	"crypto/rand"
 	"database/sql"
-	"encoding/binary"
 	"errors"
-	"fmt"
 	"testing"
 	"testing/fstest"
 	"time"
 
 	"github.com/golang-migrate/migrate/v4"
-	"github.com/jackc/pgx/v5"
 	_ "github.com/jackc/pgx/v5/stdlib"
 	"github.com/stretchr/testify/require"
 
 	"github.com/agentregistry-dev/agentregistry/pkg/registry/database"
-	"github.com/agentregistry-dev/agentregistry/pkg/registry/database/dbtest"
 )
-
-// freshDB creates a fresh per-test Postgres database. Fails when
-// PostgreSQL is unavailable.
-func freshDB(t *testing.T) string {
-	t.Helper()
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	t.Cleanup(cancel)
-
-	adminConn, adminURI := dbtest.ConnectAdmin(ctx, t)
-	defer func() { _ = adminConn.Close(ctx) }()
-
-	var randomBytes [8]byte
-	_, err := rand.Read(randomBytes[:])
-	require.NoError(t, err)
-	dbName := fmt.Sprintf("test_migrator_sp_%d", binary.BigEndian.Uint64(randomBytes[:]))
-
-	_, err = adminConn.Exec(ctx, fmt.Sprintf("CREATE DATABASE %s", dbName))
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		cleanupCtx, ccancel := context.WithTimeout(context.Background(), 10*time.Second)
-		defer ccancel()
-		c, cerr := pgx.Connect(cleanupCtx, adminURI)
-		if cerr != nil {
-			return
-		}
-		defer func() { _ = c.Close(cleanupCtx) }()
-		_, _ = c.Exec(cleanupCtx,
-			"SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = $1 AND pid <> pg_backend_pid()",
-			dbName)
-		_, _ = c.Exec(cleanupCtx, fmt.Sprintf("DROP DATABASE IF EXISTS %s", dbName))
-	})
-
-	uri, err := dbtest.DBURI(adminURI, dbName)
-	require.NoError(t, err)
-	return uri
-}
 
 // TestNewMigrator_LandsTablesInTargetSchema asserts that unqualified
 // CREATE TABLE in a migration lands in the SchemaName configured on
